exon/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
psycopg2-binary==2.9.9
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.1
requests==2.31.0
tensorflow>=2.13.0
transformers>=4.35.0
numpy>=1.24.0